split out the fs code and add placeholder for pre-1.16 fallback code
